requirements.txt
# Python dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data acquisition
yfinance>=0.2.18
requests>=2.31.0

# Technical analysis
ta>=0.11.0
pandas-ta>=0.3.14b

# Database
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.0

# Data validation
pandera>=0.15.0

# Machine Learning & RL
torch>=2.0.0
stable-baselines3[extra]>=2.0.0
gymnasium>=0.29.0
tensorboard>=2.13.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
pathlib>=1.0.1
tenacity>=8.2.0

# Testing & Quality
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.3.0
mypy>=1.3.0
flake8>=6.0.0

# Logging & Monitoring
colorlog>=6.7.0
